Fix hit golf ball with golf club
