+++ FontysSports/Assets/Prefab/Golf/GolfClub.prefab
@@ -281,7 +281,7 @@ Rigidbody:
   m_ImplicitTensor: 1
   m_UseGravity: 0
   m_IsKinematic: 1
-  m_Interpolate: 0
+  m_Interpolate: 1
   m_Constraints: 0
   m_CollisionDetection: 2
 --- !u!1 &3142441607929649598
