--- FontysSports/Assets/Prefab/Golf/GolfClubFollowManager.prefab
@@ -1,116 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1 &7985017190999010518
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 8903446141676982486}
-  - component: {fileID: 761478009980864815}
-  m_Layer: 0
-  m_Name: GolfClubFollowManager
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &8903446141676982486
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7985017190999010518}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children:
-  - {fileID: 4140494232703439280}
-  m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &761478009980864815
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7985017190999010518}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f655b55a0b51463459cda810de4487d6, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  inputAction: {fileID: 0}
-  inputManager: {fileID: 0}
-  golfClubLeft: {fileID: 0}
-  golfClubRight: {fileID: 4140494232703439280}
-  xrSimulatorActive: 0
-  forwardMultiplier: -0.02
-  upMultiplier: 0
---- !u!1001 &4587320541673940416
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    serializedVersion: 3
-    m_TransformParent: {fileID: 8903446141676982486}
-    m_Modifications:
-    - target: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3022474476115163117, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-      propertyPath: m_Name
-      value: GolfClubRight
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-    m_RemovedGameObjects: []
-    m_AddedGameObjects: []
-    m_AddedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
---- !u!4 &4140494232703439280 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 494430774154785904, guid: 68b0aeddfe1521e4e9abfec9203790e3, type: 3}
-  m_PrefabInstance: {fileID: 4587320541673940416}
-  m_PrefabAsset: {fileID: 0}
